+++ MatchingSwiftUI/MatchingSwiftUIApp.swift
@@ -11,7 +11,7 @@ import SwiftUI
 struct MatchingSwiftUIApp: App {
     var body: some Scene {
         WindowGroup {
-            ContentView()
+            HomeView()
         }
     }
 }
