+++ MatchingSwiftUI/Util/Extensions/View_extension.swift
@@ -1,0 +1,7 @@
+import SwiftUI
+
+extension View {
+    func asRoundShodow(size: CGFloat) -> some View {
+        modifier(BottomButtonModifier(size: size))
+    }
+}
