+++ MatchingSwiftUI/Util/Extensions/image_extension.swift
@@ -1,0 +1,11 @@
+import SwiftUI
+
+extension Image {
+    
+    func asTopControllButtonImage() -> some View {
+        self
+            .resizable()
+            .scaledToFill()
+            .frame(width: 30, height: 30, alignment: .center)
+    }
+}
